Build unification part three:
The original intent was to take advantage of the new consistency in
the archive filenames and their internal directory names to simplify
the version configuration needed in build.properties.

It turned out to be possible to make the fetch-third-party and
assemble-web-app targets both simpler and more flexible. The resulting
build.xml actually works with both the old archive names and paths and
the new ones.

Changes:

* Refactored the fetch-third-party and assemble-web-app targets for
  simplicity, to use fewer local temp files, and to not care about
  extraneous details of the archives. The build.xml file no longer
  refers to the component version properties.

* Simplified build.properties by removing unneeded properties
  (build-dir and onboard-flag), and by using a single version string
  for each component, based on the download URLs.

* Replaced the use of the onboard-flag property with the literal
  "manager.onboard". There is no point in the abstraction, and we
  already use the literal "manager.locked" in several places.

* Updated build.properties to point at the latest 3.0 downloads.

Other changes:

* Fixed the copyright string, typos, updated out-of-date comments, and
  moved target comments to description attributes.

Code review: http://codereview.appspot.com/6308058/